backend/requirements.txt
numpy
requests
fastapi
uvicorn
openai
python-dotenv